feat: paginate KB list endpoint and split KB form into separate page (#553)
