+++ frontend/src/Routes.tsx
@@ -29,6 +29,7 @@ import FineTuneJobDetail from "@/views/FineTune/pages/FineTuneJobDetail";
 import Tools from "@/views/Tools/Index";
 import CreateTool from "@/views/Tools/pages/CreateTool";
 import KnowledgeBase from "@/views/KnowledgeBase/Index";
+import KnowledgeBaseForm from "@/views/KnowledgeBase/pages/KnowledgeBaseForm";
 import MLModels from "@/views/MLModels/Index";
 import MLModelDetail from "@/views/MLModels/components/MLModelDetail";
 import { FeatureFlags } from "./views/Settings/pages/FeatureFlags";
@@ -331,6 +332,22 @@ export const RoutesProvider = () => {
               element: (
                 <ProtectedRoute requiredPermissions={["*", "update:knowledge_base"]}>
                   <KnowledgeBase />
+                </ProtectedRoute>
+              ),
+            },
+            {
+              path: "knowledge-base/new",
+              element: (
+                <ProtectedRoute requiredPermissions={["*", "update:knowledge_base"]}>
+                  <KnowledgeBaseForm />
+                </ProtectedRoute>
+              ),
+            },
+            {
+              path: "knowledge-base/edit/:id",
+              element: (
+                <ProtectedRoute requiredPermissions={["*", "update:knowledge_base"]}>
+                  <KnowledgeBaseForm />
                 </ProtectedRoute>
               ),
             },
